requirements.txt
click==7.1.2
decorator==4.4.2
dnspython==2.1.0
Flask==1.1.2
importlib-resources==5.1.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
netaddr==0.8.0
six==1.15.0
validators==0.18.2
Werkzeug==1.0.1
zipp==3.4.0
